backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
flask==3.0.2
python-dotenv==1.0.0
sqlalchemy==2.0.27
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai>=0.3.2
aiosqlite==0.19.0
python-dateutil==2.8.2
email-validator==2.1.0.post1
pydantic-settings==2.2.1
httpx==0.25.1 